Fix tree shaking for icons
