+++ playground/config-overrides.js
@@ -5,15 +5,28 @@ const {
   addLessLoader,
   fixBabelImports,
 } = require('customize-cra')
+
 const themeVariables = require('../src/theme-variables.json')
 
 module.exports = override(
-  // Both of these are required for using marklogic-ui-library as an ES module
+  // All three of these are required for using marklogic-ui-library as an ES module
   fixBabelImports('@marklogic/design-system',
     {
       libraryDirectory: 'es',
       camel2DashComponentName: false,
       style: true,
+    },
+  ),
+  fixBabelImports('@marklogic/design-system/MLIcon',
+    {
+      libraryDirectory: '',
+      camel2DashComponentName: false,
+      customName: function (name) {
+        return `@marklogic/design-system/es/MLIcon/${name}`
+      },
+      style: function() {
+        return "@marklogic/design-system/es/MLIcon/style"
+      },
     },
   ),
   fixBabelImports(
