+++ src/MLIcon/index.js
@@ -1,2 +1,10 @@
+import * as FontAwesomeIcons from './font-awesome-icons'
+import * as AntIcons from './ant-icons'
+
 export * from './font-awesome-icons'
 export * from './ant-icons'
+
+const MLIcon = {}
+Object.assign(MLIcon, FontAwesomeIcons, AntIcons)
+
+export default MLIcon
